Added reload support.
You can now reload the plugin with the reload() method or /ssp reload
ingame.
SSPReloadEvent was added as a notification event,  because all Group and
SSPlayer objects are not valid after a reload, and must be retrieved
again.
If you do not refresh your objects when a reload occurs, all methods in
those objects will either return null or false.